+++ github-harvester-operator/Gruntfile.js
@@ -162,7 +162,6 @@ module.exports = function (grunt) {
                     helpers: ['src/test/helpers/*.js'],
                     vendor: ['node_modules/jquery/dist/jquery.js',
                         'node_modules/jasmine-jquery/lib/jasmine-jquery.js',
-
                         'node_modules/mock-applicationmashup/lib/vendor/mockMashupPlatform.js',
                         'src/test/vendor/*.js'
                     ]